Airflow_Amazon-main/requirements.txt
# Core ETL Libraries
pandas==2.1.4
pydantic==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Airflow
apache-airflow==2.8.3

# Configuration
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Data validation
email-validator==2.1.0

# Additional utilities
openpyxl==3.10.10

# ========================================
# Team 1 - T0033: FastAPI Service Dependencies (Sprint 7)
# ========================================
# REST API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP client (for API testing)
httpx==0.26.0

# Monitoring & Observability
prometheus_client==0.19.0
